Refactor navigation and layout components; add FanNav and update Netlify configuration
