+++ app/components/common/navbar/FanNav.tsx
@@ -1,0 +1,5 @@
+const FanNav = () => {
+    return (  );
+}
+
+export default FanNav;
