+++ public/netlify.toml
@@ -1,0 +1,4 @@
+[[redirects]]
+      from = "/*"
+      to = "/index.html"
+      status = 200
